# optimize mcf