--- a/atcoder/internal_csr
+++ b/atcoder/internal_csr
@@ -1,0 +1,1 @@
+#include "atcoder/internal_csr.hpp"